Custom dc place page test fix
